+++ frontend/src/App.js
@@ -1,16 +1,11 @@
-import React, { lazy, Suspense } from 'react'
-
+import React from 'react'
 
 import './global.css'
 
 import IsNotAuthenticated from './IsNotAuthenticated'
 
 const App = () => {
-  return (
-
-   <IsNotAuthenticated/>
-
-  )
+  return <IsNotAuthenticated />
 }
 
 export default App
