+++ frontend/src/IsNotAuthenticated.js
@@ -1,10 +1,7 @@
 import React from 'react'
-import { BrowserRouter as Router, Route, Switch, Redirect } from 'react-router-dom'
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
 
-import Register from './pages/register'
-import Login from './pages/login'
-import Home from './pages/home'
-import Teste from './pages/teste'
+import { Register, Login, Home, Teste } from './pages'
 
 const IsNotAuthenticated = () => {
   return (
@@ -12,9 +9,9 @@ const IsNotAuthenticated = () => {
       <Switch>
         <Route exact path='/register' component={Register} />
         <Route exact path='/login' component={Login} />
-        <Route exact path='/' component={Home}/>
-        <Route exact path='/teste' component={Teste}/>
-        </Switch>
+        <Route exact path='/' component={Home} />
+        <Route exact path='/teste' component={Teste} />
+      </Switch>
     </Router>
   )
 }
