+++ frontend/src/components/SecondaryMenu/index.js
@@ -1,0 +1,1 @@
+export { default as SecondaryMenu } from './SecondaryMenu'
