+++ frontend/src/pages/index.js
@@ -1,0 +1,4 @@
+export * from './Register'
+export * from './Login'
+export * from './Home'
+export * from './Teste'
