+++ frontend/src/pages/teste/index.js
@@ -1,20 +1,1 @@
-import React from 'react'
-import './style.css'
-
-const teste = () => {
-    return (
-        <div className='background'>
-
-            <div className="content">
-                <div className="items"><p>1</p></div>
-                <div className="items"><p>2</p></div>
-                <div className="items"><p>3</p></div>
-                <div className="items"><p>4</p></div>
-                <div className="items"><p>5</p></div>
-                <div className="items"><p>6</p></div>
-            </div>
-        </div>
-    )
-}
-
-export default teste;
+export { default as Teste } from './Teste'
