+++ frontend/src/pages/teste/style.css
@@ -1,35 +1,29 @@
-
-
 .background {
-    display: flex;
-    width: 100vw;
-    height: 100vh;
-    background-color: rgb(20, 20, 20);   
-    justify-content: center; 
-    align-items: center;
-
+  width: 100vw;
+  height: 100vh;
+  background-color: rgb(20, 20, 20);
+  justify-content: center;
 }
 
-
 .content {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-evenly;
-    align-content: center;
-    align-items: center;
-    flex-wrap: wrap;
-    width: 500px;
-    height: 500px;
-    background-color: rgb(31, 30, 30);    
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+  align-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+  width: 500px;
+  height: 500px;
+  background-color: rgb(31, 30, 30);
 }
 
 .items {
-    display: flex;
-    color: #FFF;
-    justify-content: center;
-    align-items: center;
-    margin: 20px;
-    width: 100px;
-    height: 100px;
-    background-color: rgb(206, 53, 53);  
+  display: flex;
+  color: #fff;
+  justify-content: center;
+  align-items: center;
+  margin: 20px;
+  width: 100px;
+  height: 100px;
+  background-color: rgb(206, 53, 53);
 }
